Extract Sensor and constants to header
